Update chaps.ps1
The FuzzySecurity link referenced the GreyHatHacker post, which says that both the HKCU and HKLM registry keys must be set for the vulnerability to be present. 
Currently both the ressysele and resusrele variables check the HKLM, so the resusrele needs to be updated to check the HKCU element. Then the if statement is updated to check if both registry keys exist.